fix: fix button font weight
