+++ app/client/packages/design-system/widgets/src/components/Text/Text.tsx
@@ -4,42 +4,14 @@ import { StyledText } from "./index.styled";
 import classNames from "classnames";
 
 import type { Ref } from "react";
-import type {
-  TypographyVariant,
-  TypographyColor,
-} from "@design-system/theming";
-
-export interface TextProps {
-  /** variant of the text
-   * @default body
-   */
-  variant?: keyof typeof TypographyVariant;
-  /** color of the text
-   * @default default — sets inherit via CSS;
-   */
-  color?: keyof typeof TypographyColor;
-  /** sets the weight (or boldness) of the font
-   * @default false
-   */
-  isBold?: boolean;
-  /** sets the weight (or boldness) of the font
-   * @default false
-   */
-  isItalic?: boolean;
-  /** Sets a font that is classified as italic.
-   * @default false
-   */
-  textAlign?: "left" | "center" | "right";
-  lineClamp?: number;
-  className?: string;
-  children: React.ReactNode;
-}
+import type { TextProps } from "./types";
 
 const _Text = (props: TextProps, ref: Ref<HTMLParagraphElement>) => {
   const {
     children,
     className,
     color = "default",
+    fontWeight = 400,
     isBold = false,
     isItalic = false,
     lineClamp,
@@ -50,6 +22,7 @@ const _Text = (props: TextProps, ref: Ref<HTMLParagraphElement>) => {
 
   return (
     <StyledText
+      $fontWeight={fontWeight}
       $isBold={isBold}
       $isItalic={isItalic}
       $lineClamp={lineClamp}
